requirements.txt
pip install jira python-dateutil progressbar
pip install -U python-dotenv

python-dotenv~=0.14.0
pandas~=1.1.5
openpyxl~=3.0.5
progressbar~=2.5
python-dateutil~=2.8.1
pytz~=2020.1
telebot~=0.0.4
jira~=2.0.0